----------BUGFIX----------
The isocontours plugin was not using the isFastAlgorithmLoaded flag to
create the richmodel if not available. Otherwise if
isFastAlgorithmLoaded is true then there is no necessary to create the
python data version of Richmodel that is slower than its C-counterpart.
isFastAlgorithmLoaded is a flag that detects if the cython version of
chenhancc is available for the python interpretor.